requirements.txt
discord.py
python-dotenv
PyNaCl
ffmpeg
flask
hypercorn